refactor(storage): merge FileUpload and UploadSource into unified UploadSource type
